feat: move auth routes to ui
