+++ packages/client/src/router/AppRouter.tsx
@@ -1,12 +1,10 @@
-import { Loader } from '@codement/ui';
+import { AuthRoute, UnAuthRoute } from '@codement/ui';
 import React from 'react';
 import {
-  Redirect, Route, RouteProps, Router, Switch, useHistory
+  Redirect, Route, Router, Switch
 } from 'react-router';
-
-import { Auth } from '@codement/ui/lib/containers/Auth.container';
-import { Me } from '@codement/ui/lib/containers/Me.container';
 import { history as History } from '@codement/ui/lib/history';
+
 import { Wizard } from '../containers/Wizard.container';
 import { OnboardingPage } from '../pages/Onboarding/Onboarding.page';
 import { DashboardPage } from '../pages/Dashboard/Dashboard.page';
@@ -15,38 +13,12 @@ import { LogoutPage } from '../pages/Logout/Logout.page';
 import { routes } from './routes';
 
 
-// if (location.state?.referrer) {
-//   return location.state?.referrer;
-// }
-const getLoginRedirect = () => routes.home();
-
-const AuthRoute: React.FunctionComponent<RouteProps> = props => {
-  const { status } = Auth.useContainer();
-  const history = useHistory();
-  const { loading, called } = Me.useContainer();
-
-  if (status === 'signedOut') history.push(routes.login());
-  if ((status === 'signingIn' || status === 'verifying') || !called || loading) return <Loader />;
-
-  return <Route {...props} />;
-};
-
-
-const UnAuthRoute: React.FunctionComponent<RouteProps> = props => {
-  const { status } = Auth.useContainer();
-  const history = useHistory();
-
-  if (status === 'verifying') return <Loader />;
-  if (status === 'signedIn') history.push(getLoginRedirect());
-  return <Route {...props} />;
-};
-
 export const AppRouter = () => (
   <Router history={History}>
     <Switch>
-      <UnAuthRoute path={routes.login(false)} component={LoginPage} />
+      <UnAuthRoute routes={routes} path={routes.login(false)} component={LoginPage} />
 
-      <AuthRoute path="*">
+      <AuthRoute routes={routes} path="*">
         <Switch>
           <Route exact path={routes.home(false)} component={DashboardPage} />
           <Route exact path={routes.logout(false)} component={LogoutPage} />
